backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
gtts==2.4.0
pydub==0.25.1
python-dotenv==1.0.0 
audioop-lts==0.1.0
aiohttp==3.9.1
